# faster `isfinite` for floats