@oscardssmith 2023-01-31
Nice find! IMO this should probably be fine for all `AbstractFloat`