@giordano 2023-01-31
> It'd be nice if someone could benchmark or code_native this change on an ARM.

```julia
julia> using BenchmarkTools

julia> new_isfinite(x::AbstractFloat) = !isnan(x - x)
new_isfinite (generic function with 1 method)

julia> x = rand(100);

julia> @btime all(isfinite, $x); # before
  141.817 ns (0 allocations: 0 bytes)

julia> @btime all(new_isfinite, $x); # after
  154.057 ns (0 allocations: 0 bytes)

julia> versioninfo()
Julia Version 1.10.0-DEV.77
Commit 5da8d5f17ad (2022-11-30 11:11 UTC)
Platform Info:
  OS: Linux (aarch64-linux-gnu)
  CPU: 256 × unknown
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, thunderx2t99)
  Threads: 1 on 256 virtual cores
```
```julia
julia> code_native(isfinite, (Float64,); debuginfo=:none)
```
```as
        .text
        .file   "isfinite"
        .globl  julia_isfinite_350              // -- Begin function julia_isfinite_350
        .p2align        3
        .type   julia_isfinite_350,@function
julia_isfinite_350:                     // @julia_isfinite_350
        .cfi_startproc
// %bb.0:                               // %top
        //APP
        mrs     x8, TPIDR_EL0
        //NO_APP
        ldr     x8, [x8, #16]
        fsub    d0, d0, d0
        fcmp    d0, #0.0
        ldr     x8, [x8, #16]
        // COMPILER BARRIER
        ldr     x8, [x8, #16]
        cset    w0, eq
        ldr     xzr, [x8]
        // COMPILER BARRIER
        ret
.Lfunc_end0:
        .size   julia_isfinite_350, .Lfunc_end0-julia_isfinite_350
        .cfi_endproc
                                        // -- End function
        .section        ".note.GNU-stack","",@progbits
```
```julia
julia> code_native(new_isfinite, (Float64,); debuginfo=:none)
```
```as
        .text
        .file   "new_isfinite"
        .globl  julia_new_isfinite_372          // -- Begin function julia_new_isfinite_372
        .p2align        3
        .type   julia_new_isfinite_372,@function
julia_new_isfinite_372:                 // @julia_new_isfinite_372
        .cfi_startproc
// %bb.0:                               // %top
        //APP
        mrs     x8, TPIDR_EL0
        //NO_APP
        ldr     x8, [x8, #16]
        fsub    d0, d0, d0
        fcmp    d0, d0
        ldr     x8, [x8, #16]
        // COMPILER BARRIER
        ldr     x8, [x8, #16]
        cset    w0, vc
        ldr     xzr, [x8]
        // COMPILER BARRIER
        ret
.Lfunc_end0:
        .size   julia_new_isfinite_372, .Lfunc_end0-julia_new_isfinite_372
        .cfi_endproc
                                        // -- End function
        .section        ".note.GNU-stack","",@progbits
```
(note `cset    w0, eq` vs `cset    w0, vc`).  Results are a bit inconsistent when running repeatedly, sometimes proposed `isfinite` is on-par with current one, other times slower, but never faster.  I've got similar results on A64FX.