@mikmoore 2023-01-31
@giordano, could I trouble you to run those tests again with `@benchmark` and report the average runtimes? If a significant difference persists, the `@code_native debuginfo=:none all(isfinite,x)` for each version might also be interesting.

For x86, the inner loop of the `==0` version of `isfinite` is
```
.LBB0_4:                                # %L27
                                        # =>This Inner Loop Header: Depth=1
        cmpq    %rsi, %rcx
        je      .LBB0_7
# %bb.5:                                # %L40
                                        #   in Loop: Header=BB0_4 Depth=1
        vmovsd  (%rdx,%rsi,8), %xmm1            # xmm1 = mem[0],zero
        vsubsd  %xmm1, %xmm1, %xmm1
        incq    %rsi
        vucomisd        %xmm0, %xmm1
        jne     .LBB0_6
        jnp     .LBB0_4
```
and for the `!isnan` version it's
```
.LBB0_4:                                # %L26
                                        # =>This Inner Loop Header: Depth=1
        cmpq    %rsi, %rcx
        je      .LBB0_7
# %bb.5:                                # %L39
                                        #   in Loop: Header=BB0_4 Depth=1
        vmovsd  (%rdx,%rsi,8), %xmm0            # xmm0 = mem[0],zero
        vsubsd  %xmm0, %xmm0, %xmm0
        incq    %rsi
        vucomisd        %xmm0, %xmm0
        jnp     .LBB0_4
```
The scalar `==0` x86
```
        vsubsd  %xmm0, %xmm0, %xmm0
        vxorpd  %xmm1, %xmm1, %xmm1
        vcmpeqsd        %xmm1, %xmm0, %xmm0
        vmovq   %xmm0, %rax
        andl    $1, %eax
                                        # kill: def $al killed $al killed $rax
        retq
```
and `!isnan`
```
        vsubsd  %xmm0, %xmm0, %xmm0
        vucomisd        %xmm0, %xmm0
        setnp   %al
        retq
```
Here we see that the reducing NaN-based version saves one conditional jump over the zero-based version. In the scalar forms, it saves 2 instructions (although inlining provides the opportunity that one or both of those extra instructions might be avoided in practice) and 1 register.