@giordano 2023-02-01
> could I trouble you to run those tests again with `@benchmark` and report the average runtimes?

```
julia> using BenchmarkTools

julia> new_isfinite(x::AbstractFloat) = !isnan(x - x)
new_isfinite (generic function with 1 method)

julia> x = rand(100);

julia> @benchmark all(isfinite, $x)
BenchmarkTools.Trial: 10000 samples with 842 evaluations.
 Range (min … max):  144.524 ns … 234.796 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     145.925 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   145.858 ns ±   1.321 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

               █
  ▄▂▁▁▁▁▁▁▂▁▁▁▂█▂▂▁▁▁▁▁▁▂▁▁▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▂▂▂▂▂▁▁▂▁▂▂▂▂▂▂▂▂▂▂▂ ▂
  145 ns           Histogram: frequency by time          151 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark all(new_isfinite, $x)
BenchmarkTools.Trial: 10000 samples with 839 evaluations.
 Range (min … max):  145.875 ns … 175.952 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     147.293 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   147.270 ns ±   0.908 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▅      ▁      █▆    ▁▃                                        ▁
  █▆▃▁▁███▃▁▁▁▁▁██▁▃▃███▆▁▁▁▃▁▁▃▁▁▃▁▃▁▁▁▃▃▃▁▁▁▁▃▄▃▄▃▁▄▅▄▄▁▅▄▄▄▅ █
  146 ns        Histogram: log(frequency) by time        152 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

> If a significant difference persists, the `@code_native debuginfo=:none all(isfinite,x)` for each version might also be interesting.

```julia
julia> code_native(all, (typeof(isfinite), Vector{Float64},); debuginfo=:none)
```
```as
	.text
	.file	"all"
	.globl	julia_all_628                   // -- Begin function julia_all_628
	.p2align	3
	.type	julia_all_628,@function
julia_all_628:                          // @julia_all_628
	.cfi_startproc
// %bb.0:                               // %top
	//APP
	mrs	x8, TPIDR_EL0
	//NO_APP
	ldr	x8, [x8, #16]
	ldr	x8, [x8, #16]
	// COMPILER BARRIER
	ldr	x8, [x8, #16]
	ldr	xzr, [x8]
	// COMPILER BARRIER
	ldr	x8, [x0, #8]
	cbz	x8, .LBB0_3
// %bb.1:                               // %L17
	ldr	x9, [x0]
	ldr	d0, [x9]
	fsub	d0, d0, d0
	fcmp	d0, #0.0
	b.eq	.LBB0_4
// %bb.2:
	mov	w0, wzr
	ret
.LBB0_3:
	mov	w0, #1
	ret
.LBB0_4:                                // %L26.lr.ph
	sub	x8, x8, #1
	add	x9, x9, #8
	.p2align	2
.LBB0_5:                                // %L26
                                        // =>This Inner Loop Header: Depth=1
	cbz	x8, .LBB0_8
// %bb.6:                               // %L38
                                        //   in Loop: Header=BB0_5 Depth=1
	ldr	d0, [x9], #8
	fsub	d0, d0, d0
	sub	x8, x8, #1
	fcmp	d0, #0.0
	b.eq	.LBB0_5
// %bb.7:
	mov	w0, wzr
	ret
.LBB0_8:
	mov	w0, #1
	ret
.Lfunc_end0:
	.size	julia_all_628, .Lfunc_end0-julia_all_628
	.cfi_endproc
                                        // -- End function
	.section	".note.GNU-stack","",@progbits
```
```julia
julia> code_native(all, (typeof(new_isfinite), Vector{Float64},); debuginfo=:none)
```
```as
	.text
	.file	"all"
	.globl	julia_all_650                   // -- Begin function julia_all_650
	.p2align	3
	.type	julia_all_650,@function
julia_all_650:                          // @julia_all_650
	.cfi_startproc
// %bb.0:                               // %top
	//APP
	mrs	x8, TPIDR_EL0
	//NO_APP
	ldr	x8, [x8, #16]
	ldr	x8, [x8, #16]
	// COMPILER BARRIER
	ldr	x8, [x8, #16]
	ldr	xzr, [x8]
	// COMPILER BARRIER
	ldr	x8, [x0, #8]
	cbz	x8, .LBB0_6
// %bb.1:                               // %L17
	ldr	x9, [x0]
	ldr	d0, [x9]
	fsub	d0, d0, d0
	fcmp	d0, d0
	b.vs	.LBB0_8
// %bb.2:                               // %L25.lr.ph
	sub	x8, x8, #1
	add	x9, x9, #8
	.p2align	2
.LBB0_3:                                // %L25
                                        // =>This Inner Loop Header: Depth=1
	cbz	x8, .LBB0_7
// %bb.4:                               // %L37
                                        //   in Loop: Header=BB0_3 Depth=1
	ldr	d0, [x9], #8
	fsub	d0, d0, d0
	sub	x8, x8, #1
	fcmp	d0, d0
	b.vc	.LBB0_3
// %bb.5:
	mov	w0, wzr
	ret
.LBB0_6:
	mov	w0, #1
	ret
.LBB0_7:
	mov	w0, #1
	ret
.LBB0_8:
	mov	w0, wzr
	ret
.Lfunc_end0:
	.size	julia_all_650, .Lfunc_end0-julia_all_650
	.cfi_endproc
                                        // -- End function
	.section	".note.GNU-stack","",@progbits
```

To help comparing the two codes (I removed some extra stuff at the beginning and the end):
```diff
--- old.S	2023-02-01 23:49:49.000000000 +0000
+++ new.S	2023-02-01 23:49:53.000000000 +0000
@@ -5,36 +5,36 @@
 	ldr	xzr, [x8]
 	// COMPILER BARRIER
 	ldr	x8, [x0, #8]
-	cbz	x8, .LBB0_3
+	cbz	x8, .LBB0_6
 // %bb.1:                               // %L17
 	ldr	x9, [x0]
 	ldr	d0, [x9]
 	fsub	d0, d0, d0
-	fcmp	d0, #0.0
-	b.eq	.LBB0_4
-// %bb.2:
-	mov	w0, wzr
-	ret
-.LBB0_3:
-	mov	w0, #1
-	ret
-.LBB0_4:                                // %L26.lr.ph
+	fcmp	d0, d0
+	b.vs	.LBB0_8
+// %bb.2:                               // %L25.lr.ph
 	sub	x8, x8, #1
 	add	x9, x9, #8
 	.p2align	2
-.LBB0_5:                                // %L26
+.LBB0_3:                                // %L25
                                         // =>This Inner Loop Header: Depth=1
-	cbz	x8, .LBB0_8
-// %bb.6:                               // %L38
-                                        //   in Loop: Header=BB0_5 Depth=1
+	cbz	x8, .LBB0_7
+// %bb.4:                               // %L37
+                                        //   in Loop: Header=BB0_3 Depth=1
 	ldr	d0, [x9], #8
 	fsub	d0, d0, d0
 	sub	x8, x8, #1
-	fcmp	d0, #0.0
-	b.eq	.LBB0_5
-// %bb.7:
+	fcmp	d0, d0
+	b.vc	.LBB0_3
+// %bb.5:
 	mov	w0, wzr
 	ret
-.LBB0_8:
+.LBB0_6:
 	mov	w0, #1
 	ret
+.LBB0_7:
+	mov	w0, #1
+	ret
+.LBB0_8:
+	mov	w0, wzr
+	ret
```